+++ packages/DouYuRecorder/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Next
 
+- `recordHandle` 新增参数 `recorderType`
+
+# 1.8.0
+
 - 触发标题黑名单设定额外状态
 - 新增`debugLevel`参数，支持`none`、`basic`、`verbose`
+- 尽可能避免scdn
+- 录播姬引擎支持
 
 # 1.7.1